+++ CHANGELOG.md
@@ -24,6 +24,10 @@
   [#27](https://github.com/BottleRocketStudios/iOS-Hyperspace/issues/27)
   [#29](https://github.com/BottleRocketStudios/iOS-Hyperspace/pull/29)
 
+* Added BackendServiceHelper for public interfacing.
+  [Earl Gaspard](https://github.com/earlgaspard)
+  [#31](https://github.com/BottleRocketStudios/iOS-Hyperspace/pull/31)
+
 * Upgraded project to Xcode 9.3/Swift 4.1.
   [Tyler Milner](https://github.com/tylermilner)
   [#25](https://github.com/BottleRocketStudios/iOS-Hyperspace/issues/25)